theme

+++ src/main/java/pl/coderslab/warjees09/DrinkRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DrinkRepository extends JpaRepository<Drink, Long> {
+public interface DrinkRepository extends JpaRepository<Drink, Integer> {
 
 }
